+++ dune/dpg/applylocalfunctional_rr_impl.hh
@@ -39,7 +39,7 @@ inline static void interiorImpl(
   BlockVector<FieldVector<double,1>>
       localFunctionalVector(solutionLocalView.size());
 
-  iterateOverLocalIndexSet(
+  iterateOverLocalIndices(
     solutionLocalIndexSet,
     [&](size_t j, auto gj) {
       localFunctionalVector[j] = functionalVector[gj[0]];
@@ -142,7 +142,7 @@ faceImpl(const TestLocalView& testLocalView,
   BlockVector<FieldVector<double,1>>
       localFunctionalVector(solutionLocalView.size());
 
-  iterateOverLocalIndexSet(
+  iterateOverLocalIndices(
     solutionLocalIndexSet,
     [&](size_t j, auto gj) {
       localFunctionalVector[j]